Move Json class to App namespace
