+++ app/Insightly/Resources/OpportunityResource.php
@@ -6,7 +6,7 @@ namespace App\Insightly\Resources;
 
 use App\Domain\Integrations\Integration;
 use App\Insightly\InsightlyClient;
-use App\Insightly\Json;
+use App\Json;
 use App\Insightly\Objects\OpportunityStage;
 use App\Insightly\Serializers\OpportunitySerializer;
 use App\Insightly\Serializers\OpportunityStageSerializer;
